requirements.txt
backcall==0.1.0
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
decorator==4.4.1
elasticsearch==7.1.0
Flask==1.1.1
Flask-Cors==3.0.8
fuzzywuzzy==0.17.0
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
MarkupSafe==1.1.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.8.0
prometheus-flask-exporter==0.14.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
python-Levenshtein==0.12.0
pyyaml==5.3.1
requests==2.22.0
retrying==1.3.3
sentry-sdk[flask]==0.16.2
six==1.13.0
soupsieve==1.9.5
traitlets==4.3.3
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
